feat:Realizado o Mappers, e melhoria do projeto

+++ src/main/java/io/github/EduardoAraujoPires/sistema/ingressos/service/EventoService.java
@@ -1,42 +1,50 @@
 package io.github.EduardoAraujoPires.sistema.ingressos.service;
 
+import io.github.EduardoAraujoPires.sistema.ingressos.controller.mapper.EventoMapper;
 import io.github.EduardoAraujoPires.sistema.ingressos.dto.EventoRequestDTO;
+import io.github.EduardoAraujoPires.sistema.ingressos.dto.EventoResponseDTO;
 import io.github.EduardoAraujoPires.sistema.ingressos.expetion.idNaoEncontradoException;
 import io.github.EduardoAraujoPires.sistema.ingressos.model.Evento;
 import io.github.EduardoAraujoPires.sistema.ingressos.repository.EventoRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import java.net.URI;
 import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
 
 @Service
 @AllArgsConstructor
 public class EventoService {
 
     private final EventoRepository eventoRepository;
+    private final EventoMapper mapper;
 
     public List<Evento> findAll() {
         return eventoRepository.findAll();
     }
 
-    public Evento findByEvento(Long id) {
-        return eventoRepository.findById(id).orElseThrow(() -> new idNaoEncontradoException("Id não encontrado"));
+    public Evento findByEvento(UUID id) {
+        return eventoRepository.findById(id).orElseThrow(()-> new idNaoEncontradoException("Não foi possivel encontrar o id"));
+
     }
 
     @Transactional
-    public Evento save(EventoRequestDTO dto) {
-         Evento evento = new Evento();
-         evento.setNome(dto.getNome());
-         evento.setPreco(dto.getPreco());
-         evento.setCapacidade(dto.getCapacidade());
-         evento.setIngressosDisponiveis(dto.getCapacidade());
-
-        return eventoRepository.save(evento);
+    public ResponseEntity<?> save(EventoRequestDTO dto) {
+         Evento evento = mapper.toEntity(dto);
+        evento.setIngressosDisponiveis(dto.getCapacidade());
+        eventoRepository.save(evento);
+        URI uri = ServletUriComponentsBuilder.fromCurrentRequestUri().path("/{id}").buildAndExpand(evento.getId()).toUri();
+        return ResponseEntity.created(uri).body(evento);
     }
 
     @Transactional
-    public Evento update(Long id, EventoRequestDTO dto) {
+    public ResponseEntity<?> update(UUID id, EventoRequestDTO dto) {
         Evento eventoId = eventoRepository.findById(id).orElseThrow(() -> new idNaoEncontradoException("Id não encontrado"));
         if (id.equals(eventoId.getId())) {
             eventoId.setNome(dto.getNome());
@@ -45,16 +53,20 @@ public class EventoService {
 
             int novosIgressos = eventoId.getCapacidade() - eventoId.getIngressosDisponiveis();
             eventoId.setIngressosDisponiveis(dto.getCapacidade() - novosIgressos);
-            return eventoRepository.save(eventoId);
+            eventoRepository.save(eventoId);
+            return ResponseEntity.status(HttpStatus.OK).body(eventoId);
 
+        } else {
+            throw new idNaoEncontradoException("Id não encontrado");
         }
-        return null;
-
     }
 
     @Transactional
-    public void deleteById(Long id){
-        eventoRepository.deleteById(id);
+    public void deleteById(UUID id){
+        Evento eventoId = eventoRepository.findById(id).orElseThrow(() -> new idNaoEncontradoException("Id não encontrado"));
+        if(eventoId.getId().equals(id)) {
+            eventoRepository.deleteById(id);
+        }
     }
 
     public List<Evento> findByIngressos(Integer quantide){
